moved the utility function to a seperate file in utils dir
